Bug#14341099 - ASSERTION FAILURE: IN FILE DICT0DICT.IC LINE 1284 INNODB: FAILING ASSERTION: 0
The dict_index_t::online_status can change from: 

   COMPLETED -> ABORTED -> COMPLETED

for a CLUSTER index if the  online index build fails for some reason.  When
the state is set to ABORTED, any DML that updates the cluster index will fail
when checking the index online status due to a race in the code.

The fix is to avoid setting the interim ABORTED state for cluster indexes
if the build fails.

Remove some unnecessary #includes. Only ha_prototypes.h is required to
use DEBUG_SYNC() instrumentation.

rb://1176 Approved by Jimmy Yang