Fix for Bug#13464493 19 X HIGHER THAN EXPECTED EXEC. TIME FOR                      MYSQL_BENCH/COUNT_DISTINCT_KEY_PREFIX
The cause for the increased execution time is that the optimizer
selects to use a full index scan instead of a loose index scan for the
query. This performance regression was introduced by the fix for Bug
12540545. 

Before the fix Bug 12540545, the query would be executed as a loose
index scan using either the primary key or the secondary key depending
statistical data from the storage engine. The choice of whether to use
the primary key or the secondary key for the loose index scan is done
in get_best_group_min_max() and is based on the estimated cost for
using each of the indexes. Both estimated cost and actual execution
time is almost similar for these two access strategies, and small
variations in the statistical data determined which index that was
chosen.

After the fix for Bug 12540545, the optimizer does the same cost based
choice between whether using primary or secondary key for loose index
scan. If the choice is that the secondary key has lowest cost, then
the query is still run as a loose index scan. If the choice is that
the primary key has lowest cost, this is selected as candidate for
doing loose index scan, but a new criteria added in the fix for Bug
12540545 checks whether this is a clustered key, and if it is, then it
will not be selected for use in loose index scan. In this case
get_best_group_min_max() will return NULL and the range optimizer will
do a full index scan.

The fix for this problem is to do the test for whether the index is a
clustered primary index before it takes part in the cost based
decision for which index that gives the lowest cost estimate for loose
index scan.