BUG#11748842:CREATE TABLE: NO WARNING IF REQUESTED DUPLICATE INDEX FOR              THE SAME COLUMN.
Analysis:
--------
Creation of duplicate index is allowed since the logic to detect the
duplicate index has not been implemented.

Fix:
---
A new member 'key_is_set_by_user' has been added to struct
KEY_CREATE_INFO. This member is set when a index is created
or modified by the user. An SQL-condition of the WARNING level
is reported upon execution of such a DDL statement if all
the following conditions are satisfied:
a) Index types[BTREE|HASH] are the same.
b) Key constraint/indexes are the same.
c) The constraint/indexes fields are in the same order.
However this does not apply to the foreign key constraint.
The SQL-condition is propagated to the ERROR level
under the STRICT mode condition.