Bug#14315223 - INNODB ASSERT IN TRX0RSEG.CC FOR MYSQL_INSTALL_DB
The assert in this bug happens because ibdata1 was initialized like this;

   innodb_data_file_path=ibdata1:10M:autoextend:max:10M

InnoDB asserts when there is not enough space in the system tablespace to
allocate all pages needed at bootstrap. A maximum size of 10 Mb works in
5.1 because the initial requires size of ibdata1 was less than 10M.
But in 5.5, a change was made to allocate all 128 rollback segments at
bootstrap.  Since then, the initial size has been 10M + the default
autoextend size of 8M.  If we were to make the default initial size of
ibdata1 to be just 11M instead of 10M, it would not autoextend during
bootstrap and the initial size of each ibdata1 file would be 11M.
This patch simply changes it to 12M to make room for a few small tables
before it does the first autoextend.  This will help MTR run faster.

In 5.6, worklog 6216 changes the autoextend size from 8M to 64M.
This changes the initial size of ibdata1 from 18M in 5.5 and earlier
releases of 5.6 to 74M in the current mysql-5.6 and mysql-trunk.
So this change is especially needed in 5.6.

This patch does not fix the possibility of the assert in this bug.
But it makes it less likely that someone would use a senseless startup
setting like that by declaring a more appropriate initial size for the
system tablespace.

RB:1173 approved by Sunny and also reviewed by Jimmy.