Bug#14120360 - MYSQL MISPLANNING QUERY WITH INDEX_MERGE
The problem was for a table with three indexes idx_a(a), 
idx_b(b) and idx_ab(a,b) and a query with equality predicates
on a and b. The optimizer decided to intersect(idx_a,idx_b)
which is obviously more work than ref access on idx_ab.

In the bug, there was a strong correlation between values in 
column a and b. The optimizer does not know anything about 
correlation between columns in different indexes and uses
probability to estimate the number of intersecting rowids. One
problem was that the estimate was way off.

However, we can work around this problem by considering
indexes to merge in order of #fields they "contribute" with. 
By this we mean fields covered by an index that are not 
covered by other indexes earlier in the merge sequence.