Bug#14409088 - MAKE --EXPLICIT_DEFAULTS_FOR_TIMESTAMP SETTING                VISIBLE TO USERS
WL6292 introduced --explicit_defaults_for_timestamp server option. 
This option causes (when set) CREATE TABLE to create all TIMESTAMP 
columns as NULL with DEFAULT NULL attribute, Without this option, 
TIMESTAMP columns are NOT NULL and have implicit DEFAULT clauses. 
The old behavior is deprecated. 

The problem is that --explicit_defaults_for_timestamp option is not 
currently visible for users. This make it hard for users to make 
runtime decision that depends on knowing how the server is going to 
handle TIMESTAMP columns. This fix makes it visible to users.