Bug#14392699: WHEN DEFAULTS-FILE USED, CLIENT       CONNECTIONS HANG AFTER 'INSTALL PLUGIN'
The server uses whatever is the current working
directory at startup time to read the configuration
files etc. Then it chdir's to the data directory for
a number of valid reasons. So, when it's time to
re-read the arguments at plugin load time the relative
paths won't work as expected any more.

This is correctly detected by the plugin loading code
and an error is generated. Unfortunately when this error
is propagated up the stack one lock is not unlocked. This
causes normal server operation to be impossible.

Fixed by correctly releasing the lock in case of an error.
Also, made a minor change in the default option handling to
make sure that defaults file's path doesn't get re-evaluated.

Automated testing not really possible, since pb2 and mtr
have non-constant test directories relative layouts. Also,
its currently not feasible to start a mysqld instance with
a defaults file (with a non-absolute file name) and then use
it to execute queries that make server to re-read the
specified config file.