BUG#13804728 - 64026 ERROR CREATING RELAY LOG INFO: ERROR TRANSFERING INFORMATION
This patch is based on the patch for BUG 12334346 which was not
pushed.

The type of the worker's repository is dictated by the coordinator, i.e.
the relay log info's repository. So when the latter is changed the
former is supposed to be automatically changed too.

However, due to a bug this was not happening making replication fail.
It turns out though that the worker's repository cannot be changed
if there are gaps because this requires to read information from one
repository and write it to another and this may cause problems for
the bitmap allocated to keep track of gaps.

In a nutshell, the bitmap's size is defined by an user option and
there is no guarantee that the recorded size will match the new
size and this may lead to crashes.

To fix this bug, we could always allocate the bitmap's maximum size
and prevent any problem. However, we prefer to print a warning message
and not to change the repository's type if there may be workers with
gaps.