Bug #14320968 - FAILURES ON TRUNK- INNODB.INNODB-MULTIPLE-TABLESPACES
This test case is usually failing on a --move_file on line 425 which
is where the test script tries to start InnoDB with a condition in
which it should fail to recover and then quit right away.  The test
is very indeterminant here since it cannot know how long it takes for
InnoDB to start, fail to recover, and quit.  So it tries to clean up
the double tablespace condition often before it has been noticed.
The result file does not report any issues except on Windows, where
sometimes, MTR is trying to move the file while it is open by the
engine doing recovery.

In order to make this test determinant, the engine must use
--innodb-force-recovery=1 so that tablespaces that cannot be opened
due to finding multiple possible files are skipped in recovery.
In this way, MTR can wait for a connection, then stop the engine and
wait for disconnect before cleaning up the multiple tablespace files.

The current code in the new 5.6 branch and mysql-trunk does not handle
force-recovery very well when there are multiple possible tablespaces.
The proble that is fixed in this patch occurs when a tablespace is
skipped during REDO but opened during UNDO, which happens after the
dictionary is available. Since REDO can skip tablespaces that are
missing when using force-recovery, InnoDB must not allow those
tablespaces to be opened and used after REDO.  This is done in
fil_open_single_table_tablespace() which is done after the data
dictionary is opened.

Another problem fixed in the code related to innodb-force-recovery=ON
and multiple tablespace locations is in fil_load_table(). This may be
called for a tablespace that was not opened earlier due to
innodb-force-recovery.  It may call dict_get_and_save_data_dir_path()
which can call fil_load_table() again causing a recursive read lock on
a system table page and an assert under UNIV_SYNC_DEBUG.
So the mtr_comit() is done earlier in fil_load_table(), before checking
tablespaces.