Bug#14113704 - A STAT ON A NON EXISTING FILE CREATING A FILE RECORD IN PFS
Before this fix, the file io instrumentation for the performance schema
would instrument some operations as follows:

- For FILE_STAT, a file instrument was created before the actual call to
my_file_stat.

- For FILE_OPEN, FILE_STREAM_OPEN, a file instrument was also created
before the actual file open call.

- For FILE_DELETE, a file instrument was possibly created before the
actual call, only to be destroyed after the call.

In general, this "optimistic" approach leads to several problems.

When the actual call to open(), fopen() or stat() fails, the instrumented
file that was created needs to be cleaned up.
At best, this creates some performance overhead.
At worst, this creates instrumentation leaks when the cleanup code is missing,
as found initially with the FILE_STAT operation.

In addition, attempting to create / destroy file instrumentations in several
threads executing open / close / stat / delete on the same file concurrently
has a high risk of creating race conditions inside the file instrumentation,
even when such race conditions do not exist in the code instrumented.

The solution is to never create files optimistically in a "start" event
(before the operation), but do create file instrumentation in the "end" event
only, and only if the operation actually succeeded.

Specific changes:
- move the find_or_create_file() call from start_file_open_wait() to end_file_open_wait()
- add a name / class member to the file locker state, to hold parameters until the call
  to find_or_create_file()
- remove all file creation / deletion logic present in start_file_wait() and end_file_wait(),
  for both clarity and performances
- create new start_file_close_wait() and end_file_close_wait() instrumentation points,
  used when closing / deleting files.
- created find_file(), to be used when instrumenting FILE_DELETE: if the file is not known,
  there is no point in creating instrumentation only to remove it after the operation.