part of WL#6347 InnoDB: Index level compression stats
Protect access to the per index stats container with a mutex. We choose to
use a mutex instead of rwlock because the performance critical part is in
page_zip_compress() and page_zip_decompress() where we need write access.
So a rwlock would not increase the concurrency of the performance critical
code but it is slower to acquire than a mutex.

Hide "map<index_id_t, page_zip_stat_t>" behind a typedef.