BUG#11748012 - 34804: AFTER RE-START OF THE SCHEDULER DISABLED                       EVENTS ARE NOT DROPPED FROM DISK
BACKGROUND:
STEP 1: If the event_scheduler is ON, and we create an event, 
event scheduler checks for events present in the event queue
and starts executing. If the event has not been executed/invoked,
it is still enabled. 
STEP 2: Then we switch OFF the event scheduler.
STEP 3: When the event scheduler is switched ON after some time 
when time period of event expires, it will mark the event as 
DISABLED in Event queue and drops it from there, But it does 
not drop the event from mysql.event table. This problem is 
mentioned as part of this bug report.

ANALYSIS:
An event is marked DISABLED in the event queue when its time
period gets expired. Then it gets dropped from the event queue,
but there was no logic to drop the disabled event from 
mysql.event table. As part of the fix, logic to drop the event
from mysql.event table has been added.

FIX:
This bug is fixed by Dropping all the events with status 
DISABLED from mysql.event table.