Bug#14497827 ASSERTION IN GEOMETRY::COLLECTION_INIT_FROM_OPRESULT()
Problem: a hole of a polygon which had the same coodinates,
was converted to a point, creating this operation buffer:

1. polygon (external ring)
2. hole1   ("nornal" internal ring)
3. point   (a hole converted to a point)
4. hole2   ("normal" internal ring)

which made Geometry::collection_init_from_opresult() crash on hole2,
as a hole is not expected as an independent shape.

Fix: instead of converting to a point, completely remove such empty holes.
The fixed operation buffer corresponding to the above example is:

1. polygon (external ring)
2. hole1   ("nornal" internal ring)
3. hole2   ("normal" internal ring)