Bug#14329235 INCORRECT SEMANTICS FOR "> ALL" WHEN NULL IN SUB QUERY
In revid:norvald.ryeng@oracle.com-20120502074409-tdwrm80ndkwjq0c9 (fix
for bug Bug#13330886), MIN/MAX transformation of nullable ALL
subqueries were meant to be blocked. This bug is a case where the test
guarding against transformation fails to filter out all such
queries. This patch fixes one of two possible methods for doing the
MIN/MAX transformations and opens it up for nullable ALL subqueries.

Problem: Queries with ALL over a UNION may return an incorrect result
if the UNION result contains NULL.

Since revid:norvald.ryeng@oracle.com-20120502074409-tdwrm80ndkwjq0c9,
nullable ALL subqueries should not be transformed to MAX/MIN
queries. However, for a UNION, the test only checks the nullability of
the first SELECT, so an ALL subquery with UNION where the first SELECT
in the UNION is not nullable (but a subsequent SELECT of the UNION is)
is able to pass the test in
Item_in_subselect::single_value_transformer() meant to prevent
transformation.

The subquery is rewritten to a MAX/MIN query using an
Item_maxmin_subselect. This item uses a
select_max_min_finder_subselect object to find the maximum or minimum
value. The comparison functions in select_max_min_finder_subselect
ignores NULL values and will find the maximum or minimum non-NULL
value. This is fine for ANY subqueries, but the result of an ALL
predicate should be UNKNOWN if there are NULL values in the subquery
result, so if aggregating for ALL, select_max_min_finder_subselect
should not ignore NULLs.

Fix: Make the test for nullable subqueries handle UNIONs
correctly. Add a parameter to the constructors of
Item_maxmin_subselect and select_max_min_finder_subselect that tells
if the aggregation should ignore NULL values and check this field
during comparisons so that NULL is returned as the maximum or minimum
value if a NULL is found in the subquery result and NULL values are
not ignored.

The above change to select_max_min_finder_subselect makes it possible
to transform nullable ALL subqueries, so this patch allows
transformation of these using Item_maxmin_subselect. Transformation
using Item_sum_max/Item_sum_min is still not allowed.