Bug#14209318 ASSERTION FAILED: !TABLES || THD->LEX->IS_QUERY_TABLES_LOCKED(), FILE SQL_OPTIMI
Problem: Some queries with a HAVING clause with a function that refers
to a function in the WHERE list with a subquery as parameter cause an
assertion to be fired.

Item_row::fix_fields does not update with_subselect, so if a subquery
is wrapped in an Item_row, with_subselect and has_subquery() will be
false for all Items further up in the tree.

To find out if an item is constant or not, Item::const_item() checks
can_be_evaluated_now(), which again checks has_subquery(). Since
Item_row::has_subquery() returns an incorrect answer, items may be
taken to be constant when they are not. This may lead to evaluating an
item during resolving, even if the item is not ready to be evaluated
yet.

Fix: Set with_subselect in Item_row::fix_fields().