Bug#14287539 CRASH (ASSERT) IN CREATE_SORT_INDEX WHEN USING SQL_BIG_RESULT ON JOIN QUERY
Problem: An assertion fires for some queries with JOIN and the
SQL_BIG_RESULT hint.

Tables that are accessed in a way that guarantees they will return
zero or one rows may be treated as constants. If the join condition
contains references to such tables, the optimizer may deduce that
further tables can be treated as constants and hence change the join
type from JT_REF to JT_CONST in create_ref_for_key(). At this point,
however, it is too late to move the JOIN_TAB to the head of the list
and increment JOIN::const_tables.

Due to this, the executor will enter create_sort_index() for this
table, but since the join type has been changed from JT_REF to
JT_CONST, an assertion fires.

The assertion, and this bug, was introduced by
revid:evgeny.potemkin@oracle.com-20120531162137-cbjkcjul3c3vyldh as
part of WL#6071.

Fix: In JOIN::make_tmp_tables_info(), skip sorting if join_type is
JT_CONST.