Patch for Bug#14483430 (RECONSIDER MYSQL 5.6.6 BIND-ADDRESS BEHAVIOUR WHEN CONFIGURED AS DEFAULT *).
The bug was introduced by WL#6301 (Change default value for the 'bind-address' option).

The user visible problem was that the server didn't start with the default value
of bind-address, which is '*'. The star means that the server should choose
'::' or '0.0.0.0' depending on whether IPv6 is available on the box or not.

The IPv6 detection was implemented by calling getaddrinfo() for the IPv6-address.
It turned out, it is not enough -- getaddrinfo() might successfully return, but later
OS failed to create an IPv6-socket. This problem is platform dependent
(reproduced on CentOS 5.2).

The fix is to make the check more strict -- try to create a dummy IPv6-socket
after getaddrinfo() succeeds.

No test case for the bug since it is platform specific.