BUG#14497598: INCORRECT RANGE CONSTRUCTION FOR COMPLEX CONDITION
Range optimizer: It is possible that a SEL_TREE contains both
a) simple range predicates (in tree->keys[]) and
b) index merge range predicates (in tree->merges)

If a tree has both, they represent equally *valid* range
predicate alternatives; both will return all relevant rows 
from the table but one may return more unnecessary rows than
the other. tree_or() was not up for the task of merging 
such complex trees. The fix is to remove the index 
merge predicates (if we have both types) before OR'ing the 
trees. 

For future reference: It is possible to tree_or such 
SEL_TREEs, it's just not prioritized ATM.