Update of SPJ component in pre 7.2 branches:
The code for the SPJ block has been added to cluster-7.0 -> even though
it is not intergrated & used from SQL before version 7.2.

We have a policy of having minimal changes in codebase between the
different 7.x branches. However, most (all) SPJ fixes has been 
applied to 7.2 only. This has made it increasingly difficult to
push a patch to 7.0 and merge it up: If it contained SPJ 
changes it most certainly would encounter conflicts.

This patch intends to rectify that by aligning the 7.0 SPJ
code with the 7.2 codebase.