Fix for bug#14143553:JOB BUFFER FULL - DATA NODE CRASH (Blizzard) and duplicate Bug#13799800 NDBMTD CRASHES DURING SONY-QUERY WITH 128 PARTITIONS ON 4 NODES WITH 4 LDM EACH
The patch extends and redifines the signals GSN_DIH_SCAN_GET_NODES_REQ, _CONF and _REF.
On order to avoid generating too many signals, which breaks the 1::4 fanout rule for
signals consumed::produced these signals may now be 'long'.

Both a short and long version of the modified signals are defined.
The short signal is only used for a single fragment.
This (the short) is mainly for SUMA and BACKUP which never request info for
more than a single fragment at at a time. All modified blocks will
handle both long and short version of the signal. If a long signal
was received, the reply will also be 'long'.