Bug#13897025 - SLAVE_PARALLEL_WORKERS DEGRADES REPLICATION PERFORMANCE
The reason of degradation appeared to be two-fold.
The load of the test report contained FK-related databases with rather
aggressive load to them forcing MTS frequently switching to sequential
execution which is somewhat expensive.
The other reason turns out to be a mistake in the source code
in resetting a multiplier to the Coordinator nap. The multiplier is computed 
as a number of excessive events counting from a certain overrun level
of the Worker queue. It was not reset to zero that made the multiplier grow
over time which manifested in the observed slave execution slowdown.

Fixed with correcting the nap factor computing.
An assert is deployed to prove the correct resetting in the case of one
Worker.
In order to hunt over possible performance reports more statistics is made
to printed into error log with --log-warnings=2.