BUG#13736244: SQL_BIG_RESULT PRODUCES AN ADDITIONAL ROW               FOR A JOIN+SUBQ+GROUP BY QUERY
When a scalar subselect in the SELECT clause is also used for
grouping and the "ordering for group by" can be resolved using 
filesort on the first table, the subquery will be evaluated 
1) once for each row combination in the query during sort index
creation and 2) once for each row combination in the query 
because end_send_group() will evaluate the group fields to
see if the value changed. This is potentially a lot of 
unnecessary work. The fix is to sort using a temp table in 
such situations.