Fix for Bug#14287654 PERFORMANCE REGRESSION FOR DESCENDING REF ACCESS SCAN                      ON INDEX WHEN USING ICP
The performance regression is caused by use of ICP in combination with
ref access that reads from an index in descending order.

The server uses the function join_read_prev_same() (in
sql_executor.cc) to read records. This function uses
handler::ha_index_next() to get the next record from the table and
then checks if the returned record has the same key value as the ref
key. This works fine when ICP is not used. With ICP this produces
correct results but can cause performance issues. The cause for this
is the use of ha_index_next() which will read records from the index
until it has found a record that satisfies the pushed index condition.
Since the storage engine does not know that it should stop reading
when a record with a different key is found, it will continue reading
records from the index until it has found a qualifying record. In the
case reported in this bug the storage engine will continue scaning
the index until it has reached the beginning of the index.

The fix for this problem is to not use ICP when ref access in
descending order is used for the query.