Bug#14348858: EXTRA ROWS IN RESULT WHEN SQL_BUFFER_RESULT IS USED WITH SUBQUERY AND MYISAM
The join_init_read_record have to be used as a function for reading the
first record when a table needs to be sorted. The code that sets this
function was outside of the add_sorting_to_table function. This led to
missing initialization in the case when the first non-const table needs to be
sorted. In its turn this resulted in missing sorting step and filesort
condition not being applied.

Now add_sorting_to_table function also sets first record reading function
appropriately.