Bug#14494275 HEAP CORRUPTION IN DECIMAL2BIN WITH LOW MAX_SORT_LENGTH
The problem was that if the user sets a very small max_sort_length
this value was honored by the function sortlength() but
later ignored by make_sortkey().
This can result in buffer overrun.

Solution one: always stay within the bounds of max_sort_length
as specified by the user.
This is what we have always done, even if the documentation says
it is only done for strings/blobs.

Sorting results may be a bit strange, since we use only the first few bytes
of int/float/decimal/time.