Bug#14558324 - INNODB ASSERTION ERROR WHEN UNKNOWN OPTION ENCOUNTERED
Current directory was not set to a valid folder so Windows returned
ERROR_ACCESS_DENIED for os_file_create().  5.6 code had been changed to
report the call stack as an assert instead of just exit(1); when an unknown
error is returned from the OS.  So this patch uses exit(1) for non-debug builds
and an assert for debug builds.