Bug#14499409: CRASH IN KEY_COPY() WITH ROLLUP AND LOOSESCAN SEMIJOIN
GROUP BY WITH ROllUP will normally require a temporary table,
but if grouping on a constant (e.g., GROUP BY 1), the temporary
table is skipped and one may add sorting without a temporary
table.  Semijoin LooseScan depends on scanning in index order
to eliminate duplicates.  Hence, switching to use filesort
instead will not work.

Fix: If an index is not to be used for group by, force use of 
temporary table if LooseScan is to be used.  This prevents
sorting to be added on the same table.